Maps and tilesets now draw on selecting a room, FuncDump is gone, some fixes made to the changes in previous commits.
